Finished command line : NEED A LOT OF TESTING FANNY

+++ includes/command_line.h
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/04 14:09:42 by fcatusse          #+#    #+#             */
-/*   Updated: 2020/03/07 20:17:20 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/08 17:08:43 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ u_int8_t	sub_prompt_error(t_read *term, char sb);
 
 u_int8_t	check_caps(const char *buf, t_read *term);
 void		insert_in_buffer(const char *buf, t_read *term);
-void		insert_char_in_buffer(const char buf, t_read *term, size_t buf_index);
+void		insert_char_in_buffer(const char buf, t_read *term, ssize_t buf_index);
 void		insert_str_in_buffer(const char *d_name, t_read *term);
 
 /*
@@ -70,7 +70,7 @@ int8_t		write_history(t_core *shell);
 void		save_history(t_read *term);
 void		research_mode(t_read *term);
 int8_t		check_expansions(t_read *term);
-void		insert_content(size_t j, ssize_t i, t_read *term, char *content);
+void		insert_content(ssize_t j, ssize_t i, t_read *term, char *content);
 int64_t		call_number(t_read *term, ssize_t i);
 int64_t		callback_number(t_read *term, ssize_t i);
 int64_t		call_word(t_read *term, ssize_t i);
@@ -83,8 +83,8 @@ char		*get_home_value(t_core *shell);
 
 u_int8_t	check_subprompt(t_core *shell);
 u_int8_t	check_backslash(t_read *term, char *quote);
-u_int8_t	check_backslash_nbr(char *str, int *index);
-u_int8_t	charset_count(t_read *term, char charset, int i);
+u_int8_t	check_backslash_nbr(char *str, ssize_t *index);
+u_int8_t	charset_count(t_read *term, char charset, size_t i);
 u_int8_t	quotes_is_matching(t_read *term, char *quote);
 void		load_subprompt(char quote, t_read *term);
 u_int8_t	read_multiline(t_read *term, char *sb);
@@ -110,10 +110,10 @@ u_int8_t	read_again
 */
 
 int			my_outc(int c);
-size_t		get_width_last_line(t_read *term);
-size_t		get_width_current_line(t_read *term);
+ssize_t		get_width_last_line(t_read *term);
+ssize_t		get_width_current_line(t_read *term);
 u_int64_t	get_mask(const char *buff);
-int			is_dot(const char *d_name);
+u_int8_t	is_dot(const char *d_name);
 u_int8_t	is_tab(const char *d_name, t_read *term);
 u_int8_t	is_dir(const char *dir);
 void		split_cmd(char **to_find, t_read *term);
@@ -125,6 +125,6 @@ void		feature(t_read *term, u_int64_t value);
 */
 
 void		xtputs(char *str, int i, int (*f)(int));
-size_t		xread(int fd, char *buff, int size);
+ssize_t		xread(int fd, char *buff, size_t size);
 
 #endif
