+++ srcs/cmdline/action_keys/insert_in_buffer.c
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/09 14:37:03 by fcatusse          #+#    #+#             */
-/*   Updated: 2020/03/07 17:09:09 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/08 16:51:20 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,10 +16,10 @@
 **	To insert a char in buffer at the end of line
 */
 
-void		insert_char_in_buffer(const char buff, t_read *term, size_t buff_index)
+void		insert_char_in_buffer(const char buff, t_read *term, ssize_t buff_index)
 {
 	ft_dprintf(STDIN_FILENO, "%c", buff);
-	if (buff == NEW_LINE[0] || term->x >= (ssize_t)term->ws_col)
+	if (buff == NEW_LINE[0] || term->x >= term->ws_col)
 	{
 		if (buff == NEW_LINE[0])
 			term->x = -1;
@@ -33,11 +33,11 @@ void		insert_char_in_buffer(const char buff, t_read *term, size_t buff_index)
 	term->x_index++;
 }
 
-static void	insert_at_index(t_read *term, size_t buff_index, const char *buff)
+static void	insert_at_index(t_read *term, ssize_t buff_index, const char *buff)
 {
-	size_t	j;
+	ssize_t	j;
 
-	j = ft_strlen(term->buffer) + 1;
+	j = (ssize_t)ft_strlen(term->buffer) + 1;
 	while (--j > buff_index)
 		term->buffer[j] = term->buffer[j - 1];
 	term->buffer[buff_index] = *buff;
@@ -50,10 +50,10 @@ static void	insert_at_index(t_read *term, size_t buff_index, const char *buff)
 **			`clr_lines' => to clear all following lines from cursor
 */
 
-static void	insert_inline_char(const char *buff, t_read *term, size_t buff_index)
+static void	insert_inline_char(const char *buff, t_read *term, ssize_t buff_index)
 {
 	char	*tmp;
-	size_t	x;
+	ssize_t	x;
 
 	x = 0;
 	tmp = NULL;
@@ -63,7 +63,7 @@ static void	insert_inline_char(const char *buff, t_read *term, size_t buff_index
 	ft_strdel(&term->buffer);
 	term->buffer = ft_memalloc(BUFF_SIZE);
 	insert_str_in_buffer(tmp, term);
-	x = buff_index + term->prompt_len;
+	x = buff_index + (ssize_t)term->prompt_len;
 	while (++x < term->width)
 		move_left(buff, term);
 	ft_strdel(&tmp);
@@ -75,7 +75,7 @@ static void	insert_inline_char(const char *buff, t_read *term, size_t buff_index
 
 void		insert_str_in_buffer(const char *d_name, t_read *term)
 {
-	size_t	buff_index;
+	ssize_t	buff_index;
 	size_t	i;
 
 	i = ft_strlen(d_name);
@@ -98,17 +98,17 @@ void		insert_str_in_buffer(const char *d_name, t_read *term)
 
 void		insert_in_buffer(const char *buff, t_read *term)
 {
-	size_t	buff_index;
-	size_t	len;
-	size_t	increase_len;
+	ssize_t	buff_index;
+	ssize_t	len;
+	ssize_t	increase_len;
 
 	increase_len = 0;
 	buff_index = term->x_index - term->prompt_len;
-	len = (*buff) ? ft_strlen(buff) : 0;
+	len = (*buff) ? (ssize_t)ft_strlen(buff) : 0;
 	if (term->x_index >= BUFF_SIZE || term->x_index + len >= BUFF_SIZE)
 	{
-		increase_len = ft_strlen(term->buffer) + READ_SIZE;
-		term->buffer = ft_realloc(term->buffer, increase_len);
+		increase_len = (ssize_t)ft_strlen(term->buffer) + READ_SIZE;
+		term->buffer = ft_realloc(term->buffer, (size_t)increase_len);
 	}
 	if (len > 1)
 	{
