+++ srcs/cmdline/auto_c/auto_completion.c
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/23 13:06:10 by fcatusse          #+#    #+#             */
-/*   Updated: 2020/03/07 17:23:16 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/08 16:53:29 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ void		delete_last_cmd(const char *d_name, t_read *term)
 	tmp = NULL;
 	i = term->width - (ssize_t)term->prompt_len - (ssize_t)ft_strlen(d_name);
 	ft_strdel(&tmp);
-	tmp = ft_strsub(term->buffer, 0, i);
+	tmp = ft_strsub(term->buffer, 0, (size_t)i);
 	goto_prompt(term);
 	ft_strdel(&term->buffer);
 	term->buffer = ft_memalloc(BUFF_SIZE);
@@ -72,7 +72,7 @@ void		auto_complete_mode(t_read *term)
 	char	*to_find;
 	ssize_t	i;
 
-	i = term->x_index - (ssize_t)term->prompt_len - 1;
+	i = term->x_index - term->prompt_len - 1;
 	to_find = NULL;
 	split_cmd(&to_find, term);
 	if (term->ac > 1)
