+++ srcs/cmdline/history/exp/replace_buffer.c
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/11 04:02:43 by fcatusse          #+#    #+#             */
-/*   Updated: 2020/03/07 19:10:17 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/08 17:01:24 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,26 +36,26 @@ static void	replace_buffer(t_read *term, char *begin, char *end, char *content)
 **		To insert in buffer the content found from the hst lst
 */
 
-void		insert_content(size_t j, ssize_t i, t_read *term, char *content)
+void		insert_content(ssize_t j, ssize_t i, t_read *term, char *content)
 {
 	char	*begin;
 	char	*end;
-	size_t	inc_len;
-	size_t	buff_len;
+	ssize_t	inc_len;
+	ssize_t	buff_len;
 
 	inc_len = 0;
 	end = NULL;
 	begin = NULL;
-	buff_len = ft_strlen(term->buffer);
+	buff_len = (ssize_t)ft_strlen(term->buffer);
 	if (i >= BUFF_SIZE)
 	{
-		inc_len = ft_strlen(term->buffer) + ft_strlen((char *)content);
-		term->buffer = ft_realloc(term->buffer, inc_len);
+		inc_len = (ssize_t)ft_strlen(term->buffer) + (ssize_t)ft_strlen((char *)content);
+		term->buffer = ft_realloc(term->buffer, (size_t)inc_len);
 	}
 	if (i > 0)
-		begin = ft_strsub(term->buffer, 0, (ssize_t)i);
-	if (i + j < buff_len)
-		end = ft_strsub(term->buffer, (unsigned int)i + j, (size_t)(buff_len - (i + j)));
+		begin = ft_strsub(term->buffer, 0, (size_t)i);
+	if ((i + j) < buff_len)
+		end = ft_strsub(term->buffer, (unsigned int)(i + j), (size_t)(buff_len - (i + j)));
 	ft_strdel(&term->buffer);
 	replace_buffer(term, begin, end, content);
 }
